+++ src/search/mod.rs
@@ -1,3 +1,4 @@
+pub mod attachment_search;
 pub mod client;
 #[cfg(feature = "ssr")]
 pub mod reindex;